MLPGRAPHMODEL IMPORT/EXPORT DONE!!!!!!
Former-commit-id: 4cd0d7756561f629ec64c8e74ebed60c88093980 [formerly a12badece9e72515744403cd9ca0c01ae10e6e18]
Former-commit-id: f46599bd2e6a99a58f14861ae3f8004948a8120e

+++ src/gui/utils/ModelTestWorker.cpp
@@ -23,7 +23,7 @@ void ModelTestWorker::process() {
 				return;
 			int expected = model->getMostProbablePrediction(dataset[0][k].y.ToVector());
 			int got = model->Predict(dataset[0][k].x);
-			for (int l = 0; l < result->size(); ++l) {
+			for (size_t l = 0; l < result->size(); ++l) {
 				if (expected != l && got != l)
 					(*result)[l].tn++;
 				else if (expected == l && got == l)
